refactor advertiser panel
