+++ common/models/ad.go
@@ -9,6 +9,7 @@ type Ad struct {
 	Bid             int64 `gorm:"not null"`
 	Active          bool  `gorm:"default:true;index"`
 	TotalCost       int64 `gorm:"default:0"`
+	Website         string
 	AdvertiserID    uint
 	Advertiser      Advertiser
 	AdsInteractions []AdsInteraction
